Trait storage_client::StorageRead

pub trait StorageRead: Send + Sync {
    fn update_to_latest_ledger(
        &self, 
        client_known_version: Version, 
        request_items: Vec<RequestItem>
    ) -> Result<(Vec<ResponseItem>, LedgerInfoWithSignatures, ValidatorChangeEventWithProof, AccumulatorConsistencyProof)>;

    fn update_to_latest_ledger_async(
        &self, 
        client_known_version: Version, 
        request_items: Vec<RequestItem>
    ) -> Pin<Box<dyn Future<Output = Result<(Vec<ResponseItem>, LedgerInfoWithSignatures, ValidatorChangeEventWithProof, AccumulatorConsistencyProof)>> + Send>>;

    fn get_transactions(
        &self, 
        start_version: Version, 
        batch_size: u64, 
        ledger_version: Version, 
        fetch_events: bool
    ) -> Result<TransactionListWithProof>;

    fn get_transactions_async(
        &self, 
        start_version: Version, 
        batch_size: u64, 
        ledger_version: Version, 
        fetch_events: bool
    ) -> Pin<Box<dyn Future<Output = Result<TransactionListWithProof>> + Send>>;

    fn get_account_state_with_proof_by_version(
        &self, 
        address: AccountAddress, 
        version: Version
    ) -> Result<(Option<AccountStateBlob>, SparseMerkleProof)>;

    fn get_account_state_with_proof_by_version_async(
        &self, 
        address: AccountAddress, 
        version: Version
    ) -> Pin<Box<dyn Future<Output = Result<(Option<AccountStateBlob>, SparseMerkleProof)>> + Send>>;

    fn get_startup_info(&self) -> Result<Option<StartupInfo>>;

    fn get_startup_info_async(
        &self
    ) -> Pin<Box<dyn Future<Output = Result<Option<StartupInfo>>> + Send>>;

    fn get_latest_ledger_infos_per_epoch(
        &self, 
        start_epoch: u64
    ) -> Result<Vec<LedgerInfoWithSignatures>>;

    fn get_latest_ledger_infos_per_epoch_async(
        &self, 
        start_epoch: u64
    ) -> Pin<Box<dyn Future<Output = Result<Vec<LedgerInfoWithSignatures>>> + Send>>;
}

This trait defines interfaces to be implemented by a storage read client.

There is a 1-1 mapping between each interface provided here and a LibraDB API. A method call on this relays the query to the storage backend behind the scene which calls the corresponding LibraDB API. Both synchronized and asynchronized versions of the APIs are provided.

Required methods

fn update_to_latest_ledger(
&self,
client_known_version: Version,
request_items: Vec<RequestItem>
) -> Result<(Vec<ResponseItem>, LedgerInfoWithSignatures, ValidatorChangeEventWithProof, AccumulatorConsistencyProof)>

See LibraDB::update_to_latest_ledger.

fn update_to_latest_ledger_async(
&self,
client_known_version: Version,
request_items: Vec<RequestItem>
) -> Pin<Box<dyn Future<Output = Result<(Vec<ResponseItem>, LedgerInfoWithSignatures, ValidatorChangeEventWithProof, AccumulatorConsistencyProof)>> + Send>>

See LibraDB::update_to_latest_ledger.

update_to_latest_ledger

fn get_transactions(
&self,
start_version: Version,
batch_size: u64,
ledger_version: Version,
fetch_events: bool
) -> Result<TransactionListWithProof>

See LibraDB::get_transactions.

fn get_transactions_async(
&self,
start_version: Version,
batch_size: u64,
ledger_version: Version,
fetch_events: bool
) -> Pin<Box<dyn Future<Output = Result<TransactionListWithProof>> + Send>>

See LibraDB::get_transactions.

fn get_account_state_with_proof_by_version(
&self,
address: AccountAddress,
version: Version
) -> Result<(Option<AccountStateBlob>, SparseMerkleProof)>

See LibraDB::get_account_state_with_proof_by_version.

fn get_account_state_with_proof_by_version_async(
&self,
address: AccountAddress,
version: Version
) -> Pin<Box<dyn Future<Output = Result<(Option<AccountStateBlob>, SparseMerkleProof)>> + Send>>

See LibraDB::get_account_state_with_proof_by_version.

fn get_startup_info(&self) -> Result<Option<StartupInfo>>

See LibraDB::get_startup_info.

fn get_startup_info_async(
&self
) -> Pin<Box<dyn Future<Output = Result<Option<StartupInfo>>> + Send>>

See LibraDB::get_startup_info.

fn get_latest_ledger_infos_per_epoch(
&self,
start_epoch: u64
) -> Result<Vec<LedgerInfoWithSignatures>>

See LibraDB::get_latest_ledger_infos_per_epoch.

fn get_latest_ledger_infos_per_epoch_async(
&self,
start_epoch: u64
) -> Pin<Box<dyn Future<Output = Result<Vec<LedgerInfoWithSignatures>>> + Send>>

See LibraDB::get_latest_ledger_infos_per_epoch.

Loading content...