Trait storage_client::StorageRead

pub trait StorageRead: Send + Sync {
#[must_use]
    fn update_to_latest_ledger<'life0, 'async_trait>(
        &'life0 self, 
        client_known_version: Version, 
        request_items: Vec<RequestItem>
    ) -> Pin<Box<dyn Future<Output = Result<(Vec<ResponseItem>, LedgerInfoWithSignatures, EpochChangeProof, AccumulatorConsistencyProof)>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn get_transactions<'life0, 'async_trait>(
        &'life0 self, 
        start_version: Version, 
        batch_size: u64, 
        ledger_version: Version, 
        fetch_events: bool
    ) -> Pin<Box<dyn Future<Output = Result<TransactionListWithProof>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn get_latest_state_root<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<(Version, HashValue)>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn get_latest_account_state<'life0, 'async_trait>(
        &'life0 self, 
        address: AccountAddress
    ) -> Pin<Box<dyn Future<Output = Result<Option<AccountStateBlob>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn get_account_state_with_proof_by_version<'life0, 'async_trait>(
        &'life0 self, 
        address: AccountAddress, 
        version: Version
    ) -> Pin<Box<dyn Future<Output = Result<(Option<AccountStateBlob>, SparseMerkleProof)>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn get_startup_info<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Option<StartupInfo>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn get_epoch_change_ledger_infos<'life0, 'async_trait>(
        &'life0 self, 
        start_epoch: u64, 
        end_epoch: u64
    ) -> Pin<Box<dyn Future<Output = Result<EpochChangeProof>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn backup_account_state<'life0, 'async_trait>(
        &'life0 self, 
        version: u64
    ) -> Pin<Box<dyn Future<Output = Result<BoxStream<Result<BackupAccountStateResponse, Error>>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn get_account_state_range_proof<'life0, 'async_trait>(
        &'life0 self, 
        rightmost_key: HashValue, 
        version: Version
    ) -> Pin<Box<dyn Future<Output = Result<SparseMerkleRangeProof>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn backup_transaction<'life0, 'async_trait>(
        &'life0 self, 
        start_version: Version, 
        num_transactions: u64
    ) -> Pin<Box<dyn Future<Output = Result<BoxStream<Result<BackupTransactionResponse, Error>>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn backup_transaction_info<'life0, 'async_trait>(
        &'life0 self, 
        start_version: Version, 
        num_transaction_infos: u64
    ) -> Pin<Box<dyn Future<Output = Result<BoxStream<Result<BackupTransactionInfoResponse, Error>>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]
    fn batch_fetch_config<'life0, 'async_trait>(
        &'life0 self, 
        access_paths: Vec<AccessPath>
    ) -> Pin<Box<dyn Future<Output = Result<Vec<Vec<u8>>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

This trait defines interfaces to be implemented by a storage read client.

There is a 1-1 mapping between each interface provided here and a LibraDB API. A method call on this relays the query to the storage backend behind the scene which calls the corresponding LibraDB API. Both synchronized and asynchronized versions of the APIs are provided.

Required methods

#[must_use] fn update_to_latest_ledger<'life0, 'async_trait>(
&'life0 self,
client_known_version: Version,
request_items: Vec<RequestItem>
) -> Pin<Box<dyn Future<Output = Result<(Vec<ResponseItem>, LedgerInfoWithSignatures, EpochChangeProof, AccumulatorConsistencyProof)>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

See LibraDB::update_to_latest_ledger.

update_to_latest_ledger

#[must_use] fn get_transactions<'life0, 'async_trait>(
&'life0 self,
start_version: Version,
batch_size: u64,
ledger_version: Version,
fetch_events: bool
) -> Pin<Box<dyn Future<Output = Result<TransactionListWithProof>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

See LibraDB::get_transactions.

#[must_use] fn get_latest_state_root<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<(Version, HashValue)>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

See LibraDB::get_latest_state_root.

#[must_use] fn get_latest_account_state<'life0, 'async_trait>(
&'life0 self,
address: AccountAddress
) -> Pin<Box<dyn Future<Output = Result<Option<AccountStateBlob>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

See LibraDB::get_latest_account_state.

#[must_use] fn get_account_state_with_proof_by_version<'life0, 'async_trait>(
&'life0 self,
address: AccountAddress,
version: Version
) -> Pin<Box<dyn Future<Output = Result<(Option<AccountStateBlob>, SparseMerkleProof)>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

See LibraDB::get_account_state_with_proof_by_version.

#[must_use] fn get_startup_info<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Option<StartupInfo>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

See LibraDB::get_startup_info.

#[must_use] fn get_epoch_change_ledger_infos<'life0, 'async_trait>(
&'life0 self,
start_epoch: u64,
end_epoch: u64
) -> Pin<Box<dyn Future<Output = Result<EpochChangeProof>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

See LibraDB::get_epoch_change_ledger_infos.

#[must_use] fn backup_account_state<'life0, 'async_trait>(
&'life0 self,
version: u64
) -> Pin<Box<dyn Future<Output = Result<BoxStream<Result<BackupAccountStateResponse, Error>>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

See LibraDB::backup_account_state.

#[must_use] fn get_account_state_range_proof<'life0, 'async_trait>(
&'life0 self,
rightmost_key: HashValue,
version: Version
) -> Pin<Box<dyn Future<Output = Result<SparseMerkleRangeProof>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

See LibraDB::get_account_state_range_proof.

#[must_use] fn backup_transaction<'life0, 'async_trait>(
&'life0 self,
start_version: Version,
num_transactions: u64
) -> Pin<Box<dyn Future<Output = Result<BoxStream<Result<BackupTransactionResponse, Error>>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Gets a stream of transactions (for backup purpose).

#[must_use] fn backup_transaction_info<'life0, 'async_trait>(
&'life0 self,
start_version: Version,
num_transaction_infos: u64
) -> Pin<Box<dyn Future<Output = Result<BoxStream<Result<BackupTransactionInfoResponse, Error>>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Gets a stream of transaction infos (for backup purpose).

#[must_use] fn batch_fetch_config<'life0, 'async_trait>(
&'life0 self,
access_paths: Vec<AccessPath>
) -> Pin<Box<dyn Future<Output = Result<Vec<Vec<u8>>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns a vector of on-chain configs as serialized byte array Order of on-chain configs returned matches the order of access_path

Loading content...