Trait storage_client::StorageWrite

pub trait StorageWrite: Send + Sync {
#[must_use]
    fn save_transactions<'life0, 'async_trait>(
        &'life0 self, 
        txns_to_commit: Vec<TransactionToCommit>, 
        first_version: Version, 
        ledger_info_with_sigs: Option<LedgerInfoWithSignatures>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

This trait defines interfaces to be implemented by a storage write client.

There is a 1-1 mappings between each interface provided here and a LibraDB API. A method call on this relays the query to the storage backend behind the scene which calls the corresponding LibraDB API. Both synchronized and asynchronized versions of the APIs are provided.

Required methods

#[must_use] fn save_transactions<'life0, 'async_trait>(
&'life0 self,
txns_to_commit: Vec<TransactionToCommit>,
first_version: Version,
ledger_info_with_sigs: Option<LedgerInfoWithSignatures>
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

See LibraDB::save_transactions.

Loading content...